+++ lib/preface_laws/preface_laws.mli
@@ -17,6 +17,7 @@ module Monoid = Monoid
 
 module Meet_semilattice = Meet_semilattice
 module Bounded_meet_semilattice = Bounded_meet_semilattice
+module Join_semilattice = Join_semilattice
 
 (** {1 Functor hierarchy} *)