+++ lib/preface_specs/preface_specs.mli
@@ -21,6 +21,7 @@ module Monoid = Monoid
 
 module Meet_semilattice = Meet_semilattice
 module Bounded_meet_semilattice = Bounded_meet_semilattice
+module Join_semilattice = Join_semilattice
 
 (** {1 Functor hierarchy} *)
 
@@ -98,6 +99,7 @@ module type SEMIGROUP = Semigroup.API
 module type MONOID = Monoid.API
 module type MEET_SEMILATTICE = Meet_semilattice.API
 module type BOUNDED_MEET_SEMILATTICE = Bounded_meet_semilattice.API
+module type JOIN_SEMILATTICE = Join_semilattice.API
 module type FUNCTOR = Functor.API
 module type BIFUNCTOR = Bifunctor.API
 module type PROFUNCTOR = Profunctor.API