+++ test/preface_laws_test/misc.ml
@@ -57,6 +57,17 @@ module Ord = struct
       | Eq, Eq -> Eq
     ;;
   end)
+
+  module Join_semilattice = Preface.Make.Join_semilattice.Via_join (struct
+    type nonrec t = t
+
+    let join x y =
+      match (x, y) with
+      | Gt, _ | _, Gt -> Gt
+      | Lt, x | x, Lt -> x
+      | Eq, Eq -> Eq
+    ;;
+  end)
 end
 
 module Bool_meet_semilattice = Preface.Make.Meet_semilattice.Via_meet (struct
@@ -71,6 +82,11 @@ Preface.Make.Bounded_meet_semilattice.Via_meet_and_top (struct
 
   let top = false
   let meet x y = x && y
+
+module Bool_join_semilattice = Preface.Make.Join_semilattice.Via_join (struct
+  type t = bool
+
+  let join x y = x || y
 end)
 
 module Sum_monoid_suite = Preface.Qcheck.Monoid.Suite (Sample.Int) (Sum)
@@ -86,6 +102,12 @@ module Bool_bounded_meet_semilattice_suite =
   Preface.Qcheck.Bounded_meet_semilattice.Suite
     (Sample.Bool)
     (Bool_bounded_meet_semilattice)
+
+module Bool_join_semilattice_suite =
+  Preface.Qcheck.Join_semilattice.Suite (Sample.Bool) (Bool_join_semilattice)
+
+module Ordering_join_semilattice_suite =
+  Preface.Qcheck.Join_semilattice.Suite (Ord) (Ord.Join_semilattice)
 
 module YOCaml_profunctor =
   Preface.Qcheck.Profunctor.Suite
@@ -157,6 +179,8 @@ let cases ~count =
     ; ("Ord Meet_semilattice", Ordering_meet_semilattice_suite.tests)
     ; ( "Bool Bounded_meet_semilattice"
       , Bool_bounded_meet_semilattice_suite.tests )
+    ; ("Bool Join_semilattice", Bool_join_semilattice_suite.tests)
+    ; ("Ord Join_semilattice", Ordering_join_semilattice_suite.tests)
     ; ("YOCaml Profunctor", YOCaml_profunctor.tests)
     ; ("YOCaml Strong", YOCaml_strong.tests)
     ; ("YOCaml Choice", YOCaml_choice.tests)