+++ CHANGES.md
@@ -1,6 +1,7 @@
 ## dev (unreleased)
 
 - Add `Bounded_meet_semilattice` [**@hakimba**](https://github.com/hakimba)
+- Add `Join_semilattice` [**@hakimba**](https://github.com/hakimba)
 - Add `Meet_semilattice` [**@xvw**](https://github.com/xvw) and [**@hakimba**](https://github.com/hakimba)
 - Add `Equivalence` in `Preface.Stdlib` [**@gr-im**](https://github.com/gr-im), [**@gxvw**](https://github.com/xvw)
 - Add `Preface.Laws` and `Preface.Qcheck` to allow external users to build their own test suites for implementations outside Preface [**@gr-im**](https://github.com/gr-im), [**@gxvw**](https://github.com/xvw)